L1_o3_profile: changes needed to compile with warnings switched on
   * With one exception:  The -fno-range-check option is needed
     to have the compiler ignore integer overflow in m_convert_coadd.f90
     (literal 4294967296).  But since I plan to replace most of that
     code very soon, I'm ignoring the issue for the moment.

+++ L1_o3_profile/src/m_get_o3prof.f90
@@ -249,7 +249,7 @@ SUBROUTINE get_apriori_covar( nz, ps, zs, ozprof, ntp,  sao3)
              30., 50., 70., 100., 150., 200., 300., 500., 700., 1000.0/)
      call get_fortstd (astd(1:nref )) 
   ELSE IF (which_aperr == 2 ) THEN
-     call get_mcprof(astd(1:nref), 2) 
+     call get_mcprof(astd(1:nref), 2)
   ELSE IF (which_aperr >= 8 .and. which_aperr <=9) THEN
      call get_tbprof (astd(1:nref),2) 
   ELSE  ! IF (which_aperr == 10) THEN
@@ -1045,7 +1045,7 @@ SUBROUTINE get_mcprof(ozref, which_out)
   ! ======================
   ! Input/Output variables
   ! ======================
-  INTEGER (KIND=dp), INTENT (IN)               :: which_out ! 1=o3p, 2=std
+  INTEGER, INTENT (IN)               :: which_out ! 1=o3p, 2=std
   REAL (KIND=dp), DIMENSION(nref), INTENT(OUT) :: ozref
 
   ! ======================
